refactor: modularize search logic and enhance search UI
- Decompose monolithic `SearchViewModel` by extracting logic into specialized search strategies and utilities
- Support collapsing and expanding file groups in `CodeSearchDialog` for better result navigation
- Reorganize search-related classes into specialized sub-packages (`.code`, `.file`, `.index`, `.utils`)
- Update UI in `CodeSearchDialog` with expand/collapse icons and improved padding
- Refactor `GlobExcluder`, `IndexDatabase`, and `SnippetBuilder` into new package structures and update dependencies across the module
