+++ core/main/src/main/java/com/rk/search/utils/SnippetBuilder.kt
@@ -1,4 +1,4 @@
-package com.rk.search
+package com.rk.search.utils
 
 import android.content.Context
 import android.graphics.Typeface
